requirements.txt
streamlit
audio-recorder-streamlit
openai-whisper
deep-translator
gTTS
torch
numpy
ffmpeg-python
